[TST] move misplaced lme_results files to results dir
