+++ statsmodels/regression/tests/results/generate_lme.py
