+++ statsmodels/regression/tests/results/lme_results.R
